requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
groq>=0.4.0
openai>=1.10.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
langgraph>=0.0.20

#pip install -U langchain-community